Re-worked a lot of pack sprite to deal with expanding etc later.
Changed pack sprite so it stores the original pixmap, and applies
crop/expand etc to the current pixmap. New fns to reset the data.
Added fn to expand a pixmap, adding borders (not same as padding).
Got rid of 'foreach' which works on a copy of the list, causing bug!
Made the list widget show original pixmap, not crop/rot/expanded
version.